Lots of changes here:
-- Switched from Flask implementation to aiohttp.
	- aiohttp is much more easily asynchronous
	- Flask is usually used for web apps, so it has a lot of stuff we don't need.
		-aiohttp is likely lighter and faster as a result
	- aiohttp looks very similar to Flask in terms of syntax so the structure is exactly the same and the code looks very similar

-- Under link_analysis.py, added DB support
	- Upon request, adds data to doc_store DB

NOTE: The DB is only on my local, and I'm not sure yet exactly how to spin up a new DB each time the server runs (or check if it's in the local machine)
That will likely be a future implementation.
The schema for the DB is located on the omni-indexing channel on Slack. It was posted by Cameron and ours is currently exactly identical.
So, if you want to install postgresql and create the tables manually to test on then that will work as long as the schema is copied directly. double = real, string = varchar, string[] = varchar
Otherwise, it's just standard sql statements, so as long as you feel comfortable that they work, then you should be good. I'll check GitHub often to test new commits on my local.

TODO:
-- Add DB support to all the other sections (crawling, text_transformation, and querying)
	-- Since Ranking is accessing our DB directly, no need to add DB support to their section
	-- No doubt text_transformation will be the hardest since we are getting so much information from them
	-- We need to discuss querying since it's bad practice to send JSON in a GET. Either we could move it to a POST and just return more JSON, or we could have the docID be a URL parameter. Or something else if ther are other ideas

-- Create a way to initialize DB everytime the server runs unless it already exists on the local machine. Idk how to do that yet, but I'll look much more into that in the future.